+++ src/dsp_dab/dab-constants.h
@@ -34,7 +34,7 @@
 #ifndef __DAB_CONSTANTS
 #define __DAB_CONSTANTS
 
-#include "charsets.h"
+#include "utils/charsets.h"
 #include <complex>
 #include <limits>
 #include <map>
