+++ src/dsp_dab/mot_manager.h
@@ -26,7 +26,7 @@
 #include <map>
 #include <vector>
 
-#include "charsets.h"
+#include "utils/charsets.h"
 #include "tools.h"
